backend/requirements.txt
fastapi
uvicorn
mariadb
python-dotenv
